Fixed duplicate loading issue
The problem was that paths were specified using '../lib/' in one place and './../lib/' in another and 'require' annoyingly isn't yet smart enough to figure this out. The work around is to use File.expand_path to get the minimal form and when running SiteFuelRuntime.find_processors make sure that the processor's ruby file hasn't already been loaded.

It's quite messy.