Added .filter? in addition to #filter?
It seems highly perverted to have both class and instance methods for verifying a filter's existence. They work off different methodologies, however. #filter? tests if the instance responds to the given filter; .filter? checks if the class contains a filter_XXXX method.

This was needed for the AllProcessors test suite.